pyproject.toml
[project]
name = "calculator-agent-rl"
version = "0.1.0"
description = "This is the training code for an LLM to use a calculator."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "anthropic>=0.49.0",
    "python-dotenv>=1.1.0",
    "trl>=0.16.0",
    "verifiers",
]

[tool.uv.sources]
verifiers = { path = "../verifiers", editable = true }
